Add files via upload
Visual Cs varient, i practiced working with Console and file handling (not completed, im struggling with reading, however the code will generate a document in a specified file location. Please look at the picture included with the newest commit to date (08/09/2020)
It is line 23 of the getFood file, please change the location in quotation marks to the desired file location.